Move inference to separate module
