+++ src/bit_machine/exec.rs
@@ -47,7 +47,7 @@ impl BitMachine {
     /// the given program
     pub fn for_program<App: Application>(program: &Program<App>) -> BitMachine {
         let prog = program.root();
-        let io_width = prog.source_ty().bit_width() + prog.target_ty().bit_width();
+        let io_width = prog.source_ty().bit_width + prog.target_ty().bit_width;
         BitMachine {
             data: vec![0; (io_width + prog.extra_cells_bound + 7) / 8],
             next_frame_start: 0,
@@ -286,14 +286,14 @@ impl BitMachine {
         let mut call_stack = vec![];
         let mut iters = 0u64;
 
-        let input_width = ip.source_ty.bit_width();
+        let input_width = ip.source_ty.bit_width;
         if input_width > 0 && self.read.is_empty() {
             panic!(
                 "Pleas call `Program::input` to add an input value for this program {}",
                 ip
             );
         }
-        let output_width = ip.target_ty.bit_width();
+        let output_width = ip.target_ty.bit_width;
         if output_width > 0 {
             self.new_frame(output_width);
         }
@@ -306,12 +306,12 @@ impl BitMachine {
 
             match ip.term {
                 Term::Unit => {}
-                Term::Iden => self.copy(ip.source_ty.bit_width()),
+                Term::Iden => self.copy(ip.source_ty.bit_width),
                 Term::InjL(t) => {
                     self.write_bit(false);
                     if let FinalTypeInner::Sum(ref a, _) = ip.target_ty.ty {
-                        let aw = a.bit_width();
-                        self.skip(ip.target_ty.bit_width() - aw - 1);
+                        let aw = a.bit_width;
+                        self.skip(ip.target_ty.bit_width - aw - 1);
                         call_stack.push(CallStack::Goto(ip.index - t));
                     } else {
                         panic!("type error")
@@ -320,8 +320,8 @@ impl BitMachine {
                 Term::InjR(t) => {
                     self.write_bit(true);
                     if let FinalTypeInner::Sum(_, ref b) = ip.target_ty.ty {
-                        let bw = b.bit_width();
-                        self.skip(ip.target_ty.bit_width() - bw - 1);
+                        let bw = b.bit_width;
+                        self.skip(ip.target_ty.bit_width - bw - 1);
                         call_stack.push(CallStack::Goto(ip.index - t));
                     } else {
                         panic!("type error")
@@ -332,7 +332,7 @@ impl BitMachine {
                     call_stack.push(CallStack::Goto(ip.index - s));
                 }
                 Term::Comp(s, t) => {
-                    let size = program.nodes[ip.index - s].target_ty().bit_width();
+                    let size = program.nodes[ip.index - s].target_ty().bit_width;
                     self.new_frame(size);
 
                     call_stack.push(CallStack::DropFrame);
@@ -342,14 +342,14 @@ impl BitMachine {
                 }
                 Term::Disconnect(s, t) => {
                     // Write `t`'s CMR followed by `s` input to a new read frame
-                    let size = program.nodes[ip.index - s].source_ty().bit_width();
+                    let size = program.nodes[ip.index - s].source_ty().bit_width;
                     assert!(size >= 256);
                     self.new_frame(size);
                     self.write_bytes(program.nodes[ip.index - t].cmr().as_ref());
                     self.copy(size - 256);
                     self.move_frame();
 
-                    let s_target_size = program.nodes[ip.index - s].target_ty().bit_width();
+                    let s_target_size = program.nodes[ip.index - s].target_ty().bit_width;
                     self.new_frame(s_target_size);
                     // Then recurse. Remembering that call stack pushes are executed
                     // in reverse order:
@@ -357,8 +357,7 @@ impl BitMachine {
                     // 3. Delete the two frames we created, which have both moved to the read stack
                     call_stack.push(CallStack::DropFrame);
                     call_stack.push(CallStack::DropFrame);
-                    let b_size =
-                        s_target_size - program.nodes[ip.index - t].source_ty().bit_width();
+                    let b_size = s_target_size - program.nodes[ip.index - t].source_ty().bit_width;
                     // Back not required since we are dropping the frame anyways
                     // call_stack.push(CallStack::Back(b_size));
                     // 2. Copy the first half of `s`s output directly then execute `t` on the second half
@@ -371,7 +370,7 @@ impl BitMachine {
                 Term::Take(t) => call_stack.push(CallStack::Goto(ip.index - t)),
                 Term::Drop(t) => {
                     if let FinalTypeInner::Product(ref a, _) = ip.source_ty.ty {
-                        let aw = a.bit_width();
+                        let aw = a.bit_width;
                         self.fwd(aw);
                         call_stack.push(CallStack::Back(aw));
                         call_stack.push(CallStack::Goto(ip.index - t));
@@ -385,8 +384,8 @@ impl BitMachine {
                     let bw;
                     if let FinalTypeInner::Product(ref a, _) = ip.source_ty.ty {
                         if let FinalTypeInner::Sum(ref a, ref b) = a.ty {
-                            aw = a.bit_width();
-                            bw = b.bit_width();
+                            aw = a.bit_width;
+                            bw = b.bit_width;
                         } else {
                             panic!("type error");
                         }
