+++ src/core/types.rs
@@ -1,16 +1,6 @@
-//FIXME: Remove this later
-#![allow(dead_code)]
-
-use std::{cell::RefCell, cmp, fmt, mem, rc::Rc, sync::Arc};
-
-use crate::core::{Term, TypedNode, TypedProgram};
-use crate::jet::Application;
-use crate::merkle::cmr;
 use crate::merkle::common::{MerkleRoot, TypeMerkleRoot};
 use crate::merkle::tmr::Tmr;
-use crate::Error;
-
-use super::term::UntypedProgram;
+use std::{cell::RefCell, cmp, fmt, rc::Rc, sync::Arc};
 
 #[derive(Clone, Debug)]
 pub(crate) enum Type {
@@ -43,7 +33,7 @@ pub struct FinalType {
 }
 
 impl FinalType {
-    fn unit() -> Self {
+    pub fn unit() -> Self {
         let ty = FinalTypeInner::Unit;
 
         Self {
@@ -54,7 +44,7 @@ impl FinalType {
         }
     }
 
-    fn sum(a: Arc<Self>, b: Arc<Self>) -> Self {
+    pub fn sum(a: Arc<Self>, b: Arc<Self>) -> Self {
         let ty = FinalTypeInner::Sum(a.clone(), b.clone());
 
         Self {
@@ -69,7 +59,7 @@ impl FinalType {
         }
     }
 
-    fn prod(a: Arc<Self>, b: Arc<Self>) -> Self {
+    pub fn prod(a: Arc<Self>, b: Arc<Self>) -> Self {
         let ty = FinalTypeInner::Product(a.clone(), b.clone());
 
         Self {
@@ -95,24 +85,6 @@ impl FinalType {
     }
 }
 
-pub(crate) fn pow2_types() -> [Arc<FinalType>; 11] {
-    let word0 = Arc::new(FinalType::unit());
-    let word1 = Arc::new(FinalType::sum(Arc::clone(&word0), Arc::clone(&word0)));
-    let word2 = Arc::new(FinalType::prod(Arc::clone(&word1), Arc::clone(&word1)));
-    let word4 = Arc::new(FinalType::prod(Arc::clone(&word2), Arc::clone(&word2)));
-    let word8 = Arc::new(FinalType::prod(Arc::clone(&word4), Arc::clone(&word4)));
-    let word16 = Arc::new(FinalType::prod(Arc::clone(&word8), Arc::clone(&word8)));
-    let word32 = Arc::new(FinalType::prod(Arc::clone(&word16), Arc::clone(&word16)));
-    let word64 = Arc::new(FinalType::prod(Arc::clone(&word32), Arc::clone(&word32)));
-    let word128 = Arc::new(FinalType::prod(Arc::clone(&word64), Arc::clone(&word64)));
-    let word256 = Arc::new(FinalType::prod(Arc::clone(&word128), Arc::clone(&word128)));
-    let word512 = Arc::new(FinalType::prod(Arc::clone(&word256), Arc::clone(&word256)));
-
-    [
-        word0, word1, word2, word4, word8, word16, word32, word64, word128, word256, word512,
-    ]
-}
-
 impl fmt::Display for FinalType {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.display)
@@ -130,89 +102,6 @@ impl Eq for FinalType {}
 impl std::hash::Hash for FinalType {
     fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
         self.tmr.hash(state)
-    }
-}
-
-impl FinalType {
-    pub fn bit_width(&self) -> usize {
-        self.bit_width
-    }
-
-    fn from_var(var: RcVar) -> Result<Arc<FinalType>, Error> {
-        let var = find_root(var);
-        let mut var_borr = var.borrow_mut();
-
-        let existing_type = match var_borr.var {
-            Variable::Free => Type::Unit,
-            Variable::Bound(ref ty, ref mut occurs_check) => {
-                if *occurs_check {
-                    return Err(Error::OccursCheck);
-                }
-                *occurs_check = true;
-                ty.clone()
-            }
-            Variable::EqualTo(..) => unreachable!(),
-            Variable::Finalized(ref done) => return Ok(done.clone()),
-        };
-
-        let (sub1, sub2) = match existing_type {
-            Type::Unit => {
-                let ret = Arc::new(FinalType::unit());
-                var_borr.var = Variable::Finalized(ret.clone());
-                return Ok(ret);
-            }
-            Type::Sum(ref sub1, ref sub2) => (sub1.clone(), sub2.clone()),
-            Type::Product(ref sub1, ref sub2) => (sub1.clone(), sub2.clone()),
-        };
-        drop(var_borr);
-
-        let sub1 = find_root(sub1);
-        let sub2 = find_root(sub2);
-
-        let sub1_borr = sub1.borrow_mut();
-        let final1 = match sub1_borr.var {
-            Variable::Free => {
-                drop(sub1_borr);
-                Arc::new(FinalType::unit())
-            }
-            Variable::Bound(..) => {
-                drop(sub1_borr);
-                FinalType::from_var(sub1.clone())?
-            }
-            Variable::EqualTo(..) => unreachable!(),
-            Variable::Finalized(ref f1) => {
-                let ret = f1.clone();
-                drop(sub1_borr);
-                ret
-            }
-        };
-        // drop(sub1_borr);
-        let sub2_borr = sub2.borrow_mut();
-        let final2 = match sub2_borr.var {
-            Variable::Free => {
-                drop(sub2_borr);
-                Arc::new(FinalType::unit())
-            }
-            Variable::Bound(..) => {
-                drop(sub2_borr);
-                FinalType::from_var(sub2.clone())?
-            }
-            Variable::EqualTo(..) => unreachable!(),
-            Variable::Finalized(ref f2) => {
-                let ret = f2.clone();
-                drop(sub2_borr);
-                ret
-            }
-        };
-        // drop(sub2_borr);
-
-        let ret = match existing_type {
-            Type::Unit => unreachable!(),
-            Type::Sum(..) => Arc::new(FinalType::sum(final1, final2)),
-            Type::Product(..) => Arc::new(FinalType::prod(final1, final2)),
-        };
-        var.borrow_mut().var = Variable::Finalized(ret.clone());
-        Ok(ret)
     }
 }
 
@@ -243,8 +132,8 @@ impl fmt::Debug for Variable {
 }
 
 pub(crate) struct UnificationVar {
-    var: Variable,
-    rank: usize,
+    pub var: Variable,
+    pub rank: usize,
 }
 
 impl fmt::Debug for UnificationVar {
@@ -256,286 +145,17 @@ impl fmt::Debug for UnificationVar {
 pub(crate) type RcVar = Rc<RefCell<UnificationVar>>;
 
 impl UnificationVar {
-    fn free() -> UnificationVar {
+    pub fn free() -> UnificationVar {
         UnificationVar {
             var: Variable::Free,
             rank: 0,
         }
     }
 
-    fn concrete(ty: Type) -> UnificationVar {
+    pub fn concrete(ty: Type) -> UnificationVar {
         UnificationVar {
             var: Variable::Bound(ty, false),
             rank: 0,
         }
     }
-}
-
-fn bind(rcvar: &RcVar, ty: Type) -> Result<(), Error> {
-    // Cloning a `Variable` is cheap, as the nontrivial variants merely
-    // hold `Rc`s
-    let self_var = rcvar.borrow().var.clone();
-    match self_var {
-        Variable::Free => {
-            rcvar.borrow_mut().var = Variable::Bound(ty, false);
-            Ok(())
-        }
-        Variable::EqualTo(..) => unreachable!(
-            "Tried to bind unification variable which was not \
-             the representative of its equivalence class"
-        ),
-        Variable::Finalized(..) => unreachable!(),
-        Variable::Bound(self_ty, _) => match (self_ty, ty) {
-            (Type::Unit, Type::Unit) => Ok(()),
-            (Type::Sum(al1, al2), Type::Sum(be1, be2))
-            | (Type::Product(al1, al2), Type::Product(be1, be2)) => {
-                unify(al1, be1)?;
-                unify(al2, be2)
-            }
-            // FIXME output a sane error
-            _ => {
-                //            (a, b) => {
-                /*
-                let self_s = match a {
-                    Type::Unit => "unit",
-                    Type::Sum(..) => "sum",
-                    Type::Product(..) => "prod",
-                };
-                let b_s = match b {
-                    Type::Unit => "unit",
-                    Type::Sum(..) => "sum",
-                    Type::Product(..) => "prod",
-                };
-                */
-                Err(Error::TypeCheck)
-            }
-        },
-    }
-}
-
-fn find_root(mut node: RcVar) -> RcVar {
-    loop {
-        // Double-assignment needed for pre-NLL borrowck reasons
-        let parent = match node.borrow().var {
-            Variable::EqualTo(ref parent) => Some(parent.clone()),
-            _ => None,
-        };
-        let parent = match parent {
-            Some(x) => x,
-            _ => break node,
-        };
-
-        // Extra scope for pre-NLL borrowck reasons
-        {
-            let parent_borr = parent.borrow();
-            if let Variable::EqualTo(ref grandparent) = parent_borr.var {
-                node.borrow_mut().var = Variable::EqualTo(grandparent.clone());
-            }
-        }
-        node = parent;
-    }
-}
-
-fn unify(mut alpha: RcVar, mut beta: RcVar) -> Result<(), Error> {
-    alpha = find_root(alpha);
-    beta = find_root(beta);
-
-    // Already unified, done
-    if Rc::ptr_eq(&alpha, &beta) {
-        return Ok(());
-    }
-
-    // Adjust ranks for union-find path halving
-    let rank_ord = { alpha.borrow().rank.cmp(&beta.borrow().rank) };
-    match rank_ord {
-        cmp::Ordering::Less => mem::swap(&mut alpha, &mut beta),
-        cmp::Ordering::Equal => alpha.borrow_mut().rank += 1,
-        _ => {}
-    }
-
-    // Do the unification
-    let be_var = {
-        let mut be_borr = beta.borrow_mut();
-        mem::replace(&mut be_borr.var, Variable::EqualTo(alpha.clone()))
-    };
-    match be_var {
-        Variable::Free => {} // nothing to do
-        Variable::Bound(be_type, _) => bind(&alpha, be_type)?,
-        Variable::EqualTo(..) => unreachable!(),
-        Variable::Finalized(..) => unreachable!(),
-    }
-
-    Ok(())
-}
-
-#[derive(Clone, Debug)]
-struct UnificationArrow {
-    source: Rc<RefCell<UnificationVar>>,
-    target: Rc<RefCell<UnificationVar>>,
-}
-
-/// Attach types to all nodes in a program
-pub(crate) fn type_check<Witness, App: Application>(
-    program: UntypedProgram<Witness, App>,
-) -> Result<TypedProgram<Witness, App>, Error> {
-    let vec_nodes = program.0;
-    if vec_nodes.is_empty() {
-        return Ok(TypedProgram(vec![]));
-    }
-
-    let two_0 = Type::Unit.into_rcvar();
-    let two_1 = Type::Sum(two_0.clone(), two_0).into_rcvar();
-    let two_2 = Type::Product(two_1.clone(), two_1.clone()).into_rcvar();
-    let two_4 = Type::Product(two_2.clone(), two_2.clone()).into_rcvar();
-    let two_8 = Type::Product(two_4.clone(), two_4.clone()).into_rcvar();
-    let two_16 = Type::Product(two_8.clone(), two_8.clone()).into_rcvar();
-    let two_32 = Type::Product(two_16.clone(), two_16.clone()).into_rcvar();
-    let two_64 = Type::Product(two_32.clone(), two_32.clone()).into_rcvar();
-    let two_128 = Type::Product(two_64.clone(), two_64.clone()).into_rcvar();
-    let two_256 = Type::Product(two_128.clone(), two_128.clone()).into_rcvar();
-    // pow2s[i] = 2^(2^i)
-    let pow2s = [
-        two_1, two_2, two_4, two_8, two_16, two_32, two_64, two_128, two_256,
-    ];
-
-    let mut rcs = Vec::<Rc<UnificationArrow>>::with_capacity(vec_nodes.len());
-    let mut finals = Vec::<TypedNode<Witness, App>>::with_capacity(vec_nodes.len());
-
-    // Compute most general unifier for all types in the DAG
-    for (idx, program_node) in vec_nodes.iter().enumerate() {
-        let node = UnificationArrow {
-            source: Rc::new(RefCell::new(UnificationVar::free())),
-            target: Rc::new(RefCell::new(UnificationVar::free())),
-        };
-
-        match program_node {
-            Term::Iden => unify(node.source.clone(), node.target.clone())?,
-            Term::Unit => bind(&node.target, Type::Unit)?,
-            Term::InjL(i) => {
-                let i = idx - i;
-                unify(node.source.clone(), rcs[i].source.clone())?;
-                let target_type = Type::Sum(
-                    rcs[i].target.clone(),
-                    Rc::new(RefCell::new(UnificationVar::free())),
-                );
-                bind(&node.target, target_type)?;
-            }
-            Term::InjR(i) => {
-                let i = idx - i;
-                unify(node.source.clone(), rcs[i].source.clone())?;
-                let target_type = Type::Sum(
-                    Rc::new(RefCell::new(UnificationVar::free())),
-                    rcs[i].target.clone(),
-                );
-                bind(&node.target, target_type)?;
-            }
-            Term::Take(i) => {
-                let i = idx - i;
-                unify(node.target.clone(), rcs[i].target.clone())?;
-                let target_type = Type::Product(
-                    rcs[i].source.clone(),
-                    Rc::new(RefCell::new(UnificationVar::free())),
-                );
-                bind(&node.source, target_type)?;
-            }
-            Term::Drop(i) => {
-                let i = idx - i;
-                unify(node.target.clone(), rcs[i].target.clone())?;
-                let target_type = Type::Product(
-                    Rc::new(RefCell::new(UnificationVar::free())),
-                    rcs[i].source.clone(),
-                );
-                bind(&node.source, target_type)?;
-            }
-            Term::Comp(i, j) => {
-                let (i, j) = (idx - i, idx - j);
-                unify(node.source.clone(), rcs[i].source.clone())?;
-                unify(rcs[i].target.clone(), rcs[j].source.clone())?;
-                unify(node.target.clone(), rcs[j].target.clone())?;
-            }
-            Term::Case(i, j) | Term::AssertL(i, j) | Term::AssertR(i, j) => {
-                let (i, j) = (idx - i, idx - j);
-                let var1 = Rc::new(RefCell::new(UnificationVar::free()));
-                let var2 = Rc::new(RefCell::new(UnificationVar::free()));
-                let var3 = Rc::new(RefCell::new(UnificationVar::free()));
-
-                let sum12_ty = Type::Sum(var1.clone(), var2.clone());
-                let sum12_var = Rc::new(RefCell::new(UnificationVar::free()));
-                bind(&sum12_var, sum12_ty)?;
-
-                let source_ty = Type::Product(sum12_var, var3.clone());
-                bind(&node.source, source_ty)?;
-                if let Term::Hidden(..) = vec_nodes[i] {
-                } else {
-                    bind(
-                        &find_root(rcs[i].source.clone()),
-                        Type::Product(var1.clone(), var3.clone()),
-                    )?;
-                    unify(node.target.clone(), rcs[i].target.clone())?;
-                }
-                if let Term::Hidden(..) = vec_nodes[j] {
-                } else {
-                    bind(
-                        &find_root(rcs[j].source.clone()),
-                        Type::Product(var2.clone(), var3.clone()),
-                    )?;
-                    unify(node.target.clone(), rcs[j].target.clone())?;
-                }
-            }
-            Term::Pair(i, j) => {
-                let (i, j) = (idx - i, idx - j);
-                unify(node.source.clone(), rcs[i].source.clone())?;
-                unify(node.source.clone(), rcs[j].source.clone())?;
-                bind(
-                    &node.target,
-                    Type::Product(rcs[i].target.clone(), rcs[j].target.clone()),
-                )?;
-            }
-            Term::Disconnect(i, j) => {
-                let (i, j) = (idx - i, idx - j);
-                // See chapter 6 (Delegation) of TR
-                // Be careful, this order changed! https://github.com/ElementsProject/simplicity/pull/46
-                let var_a = Rc::new(RefCell::new(UnificationVar::free()));
-                let var_b = Rc::new(RefCell::new(UnificationVar::free()));
-                let var_c = Rc::new(RefCell::new(UnificationVar::free()));
-                let var_d = Rc::new(RefCell::new(UnificationVar::free()));
-
-                let s_source = Type::Product(pow2s[8].clone(), var_a.clone()).into_rcvar();
-                let s_target = Type::Product(var_b.clone(), var_c.clone()).into_rcvar();
-                unify(rcs[i].source.clone(), s_source)?;
-                unify(rcs[i].target.clone(), s_target)?;
-
-                let node_target = Type::Product(var_b, var_d.clone()).into_rcvar();
-                unify(node.source.clone(), var_a)?;
-                unify(node.target.clone(), node_target)?;
-
-                unify(rcs[j].source.clone(), var_c)?;
-                unify(rcs[j].target.clone(), var_d)?;
-            }
-            Term::Jet(jet) => {
-                bind(&node.source, jet.source_ty.to_type(&pow2s[..]))?;
-
-                bind(&node.target, jet.target_ty.to_type(&pow2s[..]))?;
-            }
-            Term::Witness(..) | Term::Hidden(..) | Term::Fail(..) => {
-                // No type constraints
-            }
-        };
-        rcs.push(Rc::new(node));
-        // dbg!(&rcs);
-    }
-
-    // Finalize, setting all unconstrained types to `Unit` and doing the
-    // occurs check. (All the magic happens inside `FinalType::from_var`.)
-    for (index, term) in vec_nodes.into_iter().enumerate() {
-        finals.push(TypedNode {
-            cmr: cmr::compute_cmr(&finals, &term, index),
-            term,
-            source_ty: FinalType::from_var(rcs[index].source.clone())?,
-            target_ty: FinalType::from_var(rcs[index].target.clone())?,
-            index,
-        });
-    }
-
-    Ok(TypedProgram(finals))
 }
